requirements.txt
pyserial>=3.5
numpy>=1.20.0
matplotlib>=3.3.0